+++ pretty_help/__init__.py
@@ -1,4 +1,4 @@
-from .menu import *
+from .emoji_menu import *
 from .pretty_help import *
 
-__version__ = "1.3.3"
+__version__ = "2.0.0a"
